% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jct_fetch.R
\name{esac_fetch}
\alias{esac_fetch}
\title{Get ESAC Data}
\usage{
esac_fetch(
  data_url = "https://keeper.mpdl.mpg.de/f/7fbb5edd24ab4c5ca157/?dl=1"
)
}
\arguments{
\item{data_url}{URL of ESAC registry dump}
}
\description{
Get ESAC Data
}
