% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jct_fetch.R
\name{jct_fetch}
\alias{jct_fetch}
\title{Helper function to retrieve and parse Transformative Agreement Data}
\usage{
jct_fetch(esac_id = NULL, data_url = NULL)
}
\arguments{
\item{esac_id}{Transformative Agreement ID issued by ESAC}

\item{data_url}{URL to Google Spreadsheet}
}
\description{
Helper function to retrieve and parse Transformative Agreement Data
}
